requirements.txt
langchain==0.0.216
boto3== 1.34.97
python-dotenv
streamlit==1.31.1
openai==0.28.1